requirements.txt
MouseInfo==0.1.3
mysql-connector==2.2.9
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt5-stubs==5.15.6.0
PyQt5Designer==5.14.1
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
